refactured most of my code to make it work properly

+++ src/js/api/post/delete.js
@@ -1,16 +1,26 @@
 import { API_SOCIAL_POSTS } from "../constants";
+import { headers } from "../headers";
 
-export async function deletePost(id) {
+export async function deletePost() {
   try {
-    const response = await fetch(`${API_SOCIAL_POSTS}/${id}`, {
-      method: "DELETE",
-      headers: headers(),
-    });
+    const findId = new URLSearchParams(window.location.search).get("id");
+    const userConfirmed = confirm("Do you wish to delete this post?");
+
+    let response;
+
+    if (userConfirmed) {
+      response = await fetch(`${API_SOCIAL_POSTS}/${findId}`, {
+        method: "DELETE",
+        headers: headers(),
+      });
+    }
 
     if (response.ok) {
+      window.location.href = "/";
       return true;
     } else {
-      throw new Error("Failed to delete post");
+      console.error("Failed to delete post:", response.status, response.statusText);
+      return false;
     }
   } catch (error) {
     console.error(error);
